Move to HiveParquetSuite
